Feature/add transaction list page

+++ lib/widgets/responsible_chips.dart
@@ -1,6 +1,8 @@
 import 'package:flutter/material.dart' hide ProgressIndicator;
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:monster_finances/data/database/entities/account_responsible.dart';
+import 'package:monster_finances/data/database/entities/transaction.dart';
+import 'package:monster_finances/providers/current_transaction_provider.dart';
 import 'package:monster_finances/providers/last_responsible_selected_provider.dart';
 import 'package:monster_finances/providers/responsible_list_provider.dart';
 import 'package:monster_finances/widgets/custom_chips_input.dart';
@@ -9,13 +11,39 @@ import 'package:monster_finances/widgets/progress_indicator.dart';
 
 class ResponsibleChips extends HookConsumerWidget {
   const ResponsibleChips({Key? key}) : super(key: key);
+  static const String deleteConfirmationKey = "CONFIRMED";
 
   @override
   Widget build(BuildContext context, WidgetRef ref) {
+    final Transaction? currentTransaction =
+        ref.watch(currentTransactionProvider);
+
     final AsyncValue<List<AccountResponsible>> responsibleList =
         ref.watch(responsibleListProvider);
     final AsyncValue<AccountResponsible?> lastResponsibleSelected =
         ref.watch(lastResponsibleSelectedProvider);
+
+    _deleteDialog() {
+      return AlertDialog(
+        title: const Text(
+            'This responsible is linked to other transactions or accounts'),
+        content: const Text(
+            'Deleting this responsible will delete it for all linked transactions and accounts. Are you sure you want to continue?'),
+        actions: <Widget>[
+          TextButton(
+            onPressed: () => Navigator.pop(context),
+            child: const Text('Cancel'),
+          ),
+          TextButton(
+            style: TextButton.styleFrom(
+              primary: Colors.red,
+            ),
+            onPressed: () => Navigator.pop(context, deleteConfirmationKey),
+            child: const Text('Delete'),
+          ),
+        ],
+      );
+    }
 
     _buildField(List<AccountResponsible> responsibleList) {
       return CustomChipsInput(
@@ -38,17 +66,30 @@ class ResponsibleChips extends HookConsumerWidget {
           return InputChip(
             key: ObjectKey(responsible.id),
             label: Text(responsible.name),
-            selected: lastResponsibleSelected.valueOrNull?.id == responsible.id,
+            selected: currentTransaction?.responsible.targetId ==
+                    responsible.id ||
+                (currentTransaction == null &&
+                    lastResponsibleSelected.valueOrNull?.id == responsible.id),
             onSelected: (bool selected) {
               ref
                   .read(lastResponsibleSelectedNotifierProvider.notifier)
                   .update(responsible);
             },
-            onDeleted: () {
-              ref
-                  .read(responsibleListNotifierProvider.notifier)
-                  .remove(ref, responsible);
-              state.deleteChip(responsible);
+            onDeleted: () async {
+              var shouldDelete = true;
+              if (responsible.transactions.isNotEmpty ||
+                  responsible.accounts.isNotEmpty) {
+                final result = await showDialog(
+                    context: context, builder: (context) => _deleteDialog());
+                shouldDelete = result == deleteConfirmationKey;
+              }
+
+              if (shouldDelete) {
+                ref
+                    .read(responsibleListNotifierProvider.notifier)
+                    .remove(ref, responsible);
+                state.deleteChip(responsible);
+              }
             },
             materialTapTargetSize: MaterialTapTargetSize.shrinkWrap,
           );
